I am ready to begin work for the day for real.
I am not entirely satisfied with how the design came out. All in all, I think I might be relying on reflection a bit too much and it is making the design worse off as a result.

I might be able to get more simplicity if I picked a more OO design. The biggest problem is that initializing both the weights and the separately streams would be such a massive pain in the ass.

For matmult_stream, I'd be better off just doing the weights and the streams at once. This was my first idea, but I had to complicate the design by letting streams potentially work on multiple inputs.

I always do this kind of thing - instead of picking a simple choice, I really like to search around.

`matmult_stream` also has a bug where I forgot to wait on the original stream.

I said this function would be a pain in the ass and I was not wrong.

In truth all the designs have something going for them. K-FAC has some advantages over CurveBall and the backward pass method I discovered has some advantages over doing it during optimization pass.

It always an endless series of tradeoffs.

...

I've though about it for half an hour now. Let me rebuild the function from scratch with the design that I have in mind.