```     inl matmult_stream l s =         inl get_dims {data weight} =             inl get_dim = function                 | {T} -> T.dim |> inl b,a -> a,b                 | T -> T.dim             inl (b,_),(_,a) = Tuple.map get_dim (data, weight)             b,a
        inl init {d with data weight streams=l,r} =
            inl dim = get_dims {data weight}
            inl s = s.data_add {stream=l}
            {d with out = s.CudaTensor.create {elem_type=float; dim} |> dr s}

        inl run {out data weight streams=l,r} =
            l.wait_on s.data.stream
            inl s = s.data_add {stream=l}

            inl f = function {T} -> T, .T | nT -> nT, .nT
            inl (A,TA),(B,TB) = f data, f weight
            s.CudaBlas.gemm' TA TB one (primal A) (primal B) zero (primal out)

        inl bck {out data weight streams=l,r} =
            inl f' = function {T} -> T, .nT | nT -> nT, .T
            inl (A,TA),(B,TB) = f' data, f' weight
            inl out = adjoint out
            on_non_nil (inl A ->
                l.wait_on s.data.stream
                inl s = s.data_add {stream=l}
                match TA with
                | .T -> s.CudaBlas.gemm' .nT TB one out (primal B) one A
                | .nT -> s.CudaBlas.gemm' .nT TB one (primal B) out one A
                ) (adjoint A)
            on_non_nil (inl B ->
                r.wait_on s.data.stream
                inl s = s.data_add {stream=r}
                match TB with
                | .T -> s.CudaBlas.gemm' TA .nT one (primal A) out one B
                | .nT -> s.CudaBlas.gemm' TA .nT one out (primal A) one B
                ) (adjoint B)

        inl l = Liple.map init l
        Liple.iter run l
        Liple.iter (inl {stream=l,r} -> s.data.stream.wait_on l) l
        {
        out=Liple.map (inl {out} -> out) l
        bck=met _ ->
            Liple.iter bck l
            Liple.iter (inl {stream=x} -> Tuple.iter s.data.stream.wait_on x) l
        }
```

How about this?

I think it would be fine if I leave the matmult variant in the library. That is the most elegant solution. Rather than having one super generic variant that has a lot of mental overhead, it is fine if I have two for different purposes.

It is amazing now that I had some time to think about it how different the conclusions I reach are. At 5:10pm yesterday I actually wanted to continue straight to testing, but I listened to my cautious side. This is the reward I get for being lazy. Laziness exists for a reason. It is a signal that tells me it is time to recharge.

What I had yesterday was not even correct and I did not know it.

Ok, now that I have this, I can seriously consider testing. Let me make a copy of the mi layer and then I will try plugging it in. I'll have to make a proper initializer for the weights for this. Let me just sketch it out first. It won't be a big deal either way.