"4:10pm.
```
The learning rate is 2 ** -11.00083
Starting timing for: iteration 0
The time was 00:00:05.0832945 for: iteration 0
Training: 2.61636078508352
The learning rate is 2 ** -11.00083
Starting timing for: iteration 1
The time was 00:00:04.9648178 for: iteration 1
Training: 2.04905539092279
The learning rate is 2 ** -11.00083
Starting timing for: iteration 2
The time was 00:00:04.9701629 for: iteration 2
Training: 1.88799821561449
The learning rate is 2 ** -11.00083
Starting timing for: iteration 3
The time was 00:00:05.1632124 for: iteration 3
Training: 1.79801253353328
The learning rate is 2 ** -11.00083
Starting timing for: iteration 4
The time was 00:00:05.0730457 for: iteration 4
Training: 1.74178054167971
```

Nothing. I have no idea why this happens. Well, it is fine.

4:15pm. Forget the performance issue. It is just the GPU being finicky. I've done what I set out to do. With this the streaming matrix multiply is done in the most elegant manner possible and I am getting the desired speedup from it. Maybe in standard RNNs it would not be so much of a deal, but there is no reason not to use it in MIRNNs. This means it will also be convenient in plastic RNNs.

Apart from PRONG the MIRNN is complete.

The first 3 bullet points are mostly done - the next thing I need to do would be the tensor views and the optimized LSTM.

Ah, yes, I need to modify the streaming multiply in order to allow module semantics. Let me do so.

I think I will call it a day after that. I am not sure I feel like starting work on the LSTM immediately "