"4:30pm.
```
The learning rate is 2 ** -11.00083
Starting timing for: iteration 0
The time was 00:00:05.0625234 for: iteration 0
Training: 2.61776821685078
The learning rate is 2 ** -11.00083
Starting timing for: iteration 1
The time was 00:00:04.8652734 for: iteration 1
Training: 2.04968601714491
The learning rate is 2 ** -11.00083
Starting timing for: iteration 2
The time was 00:00:04.8238002 for: iteration 2
Training: 1.88748146557961
The learning rate is 2 ** -11.00083
Starting timing for: iteration 3
The time was 00:00:04.9784111 for: iteration 3
Training: 1.79772181587759
The learning rate is 2 ** -11.00083
Starting timing for: iteration 4
The time was 00:00:05.0135339 for: iteration 4
Training: 1.73889438310802
```

Ah, it hit 4.8. Yeah, the performance of this is finicky as one could expect. At any rate, I've modified the function so it uses the module iteration.

```
                    inm {input state} =
                        matmult_stream {
                            input={(weights.input) with data=input}
                            state={(weights.state) with data=out'}
                            }
```

This allows me to implement the thing as so.

I suppose I could also do it like this.

```
                    inm {input state} =
                        module_map (inl k data -> {(weights k) with data}) {input state=out'}
                        |> matmult_stream
```

...But the first version is the most readable. Although, to be honest, the original version was not bad either.

```
inm input, state = matmult_stream ({(weights.input) with data=input}, {(weights.state) with data=out'})
```

Let me go with it then."