I spent most of my time studying the EKFAC paper and now I understand it in its entirety. It is quite a brilliant paper, but as I'll be using long sequences with an inversion after each episode, I think the regular KFAC will suffice as it is.
Online Natural Gradient as a Kalman Filter
http://www.yann-ollivier.org/rech/publs/natkal.pdf

Training Multilayer Perceptrons With The Extended Kalman Algorithm
https://pdfs.semanticscholar.org/c4ea/370b7261f2a2bf3a9339cedb1ab1de348301.pdf

Now I feel like studying this.

Using network to update other networks. Maybe it is possible to go further with this somehow. I said that it might turn out that the best way of doing natural gradients would be at some point to move to RNNs and have them do the updates. This could potentially solve the credit assignment problem. To think that there was work on this 30 years ago.

Yeah, I know, I know - I was supposed to be working on the LSTM, but I got the bug for reading ML papers.

https://arxiv.org/abs/1706.03662
Practical Gauss-Newton Optimisation for Deep Learning

It occurred to me that the GN in KFGN might stand for Gauss Newton and found this. Let me add it to the mix.