I am obsessed today and not with the LSTM
"Haykin S.-Kalman filtering and neural networks (2001)"

I found this book on Genesis and I feel like reading up on it. Maybe this material will bring me close to understanding what those recurrent and tridiagonal K-FAC approximations are doing.

I feel like there is a lot to be mined when it comes to the optimization stage.

I think I'll spend a day or two just studying this book.

What I like about the Kalman filter connection is that they can be trained themselves using backprop. There was a paper from Abeel's group that compared them to LSTMs.

There exists a KF - RNN connection that I need to attempt to bridge.

This is a good time to investigate this.