Ah, what a waste. I should not have studied so much, but I'll be able to salvage the day somehow. The KF book is useless, but even though I am not going to use it, I understand the KFGN algorithm. I think I'll just stick with KFAC as the reasons to switch are not convincing enough - the improvements are just too marginal compared to the implementation difficulties.
There is an experiment I should try at some point - evaluate the input and gradient covariances as one. It would force me to implement buffers, but it would be worth it should I want to. It would also give a way of naturalizing the recurrent matrices, though I am not sure if it would work.

...Agh, I just can't reach it. There should be more here, but I do not have a way of reasoning in these spaces. All I can do is stumble along.

Forget about this.

Yesterday, I really got distracted by the EKFAC paper. It was the one that started this whole thing.

Right now I will take a break and after I am done, I'll slowly get back to work on the LSTM. Forget merging the optimization passes with the main ones, that will be something for the future when the unsupervised learning puzzle is fully cracked. Forget about CurveBall too, I've been considering replacing KFAC with it, but considering that BN improves upon it, there is no way that it will be better. I knew about that, but forgot about it which is why I went off into dreams for the past few days. Well, now I know why I should avoid it.

The optimization pass will have to be something transcendent after all - there is a certain equitable quality to it where all models are equal in the eyes of the optimizer. Last night I got really into it thinking about how would I replace backprop in order to allow long term memories. There is no way that could happen if the gradients for each layer are naturalized to have the identity covariance.

Maybe it is fine if it remains like this? Maybe I just need to accept backprop for what it is.

Maybe the future does belong to the work spearheaded by Miconi where backprop is used to support unsupervised learning algorithms. That should become known once unsupervised goal formation gets cracked.