I am completely dry.
```
        inl tensor_view d dim s =
            inl init view tns =
                inl rec loop apply =
                    inl next = function
                        | {}, {} ->
                            inl rec loop = function
                                | {from near_to}, () -> ()
                                | {}, {} ->
                                    module_map
                        | (from: int64), () -> loop (() :: apply)
                        | (from: int64), _ -> error_type "Ranges must have () as index into the initializer."

            inl tns = Struct.map' (inl _ -> s.CudaTensor.create_view {dim elem_type=float}) d.init |> heap
            function
            | .data -> tns
            | .init -> Struct.iter2' (inl view tns -> init (view ()) tns) d.init tns
            | .save stream -> Struct.iter2' (inl f tns -> f stream tns s) d.save tns
            | .load stream -> Struct.iter2' (inl f tns -> f stream tns s) d.load tns
```

I honestly have no idea how to do the loop and have no energy to push my mind into the creative zone which I would need to deal with this problem.

I did not anticipate that iterating over the tensor dimensions would be this much of a problem. It seems easier than it is at first glance. I thought since I already did this while creating the tensor views I could do it, but there I was rewriting the tensor directly. Now I need to go beyond that and I am not sure how.

Let me step away from the screen for a while to think about it and then I will call it a day.

Getting into the zone involves slowly and painstaikingly feeling out the area. It will push my subconscious towards working out the solution overnight.