```         inl tensor_view d dim s =             inl init init tns =                 inl dim = tns.dim                 inl tns = tns.basic                 inl f x next s =                     match x with                     | {} ->                         inl rec loop init x =                             match x with                             | {from near_to} -> next {s with apply=x :: self; init}                             | {} -> module_map (inl k x -> loop init.k x; ()) x                         loop s.init x                     | from -> next {s with apply=() :: self}
                inl finally {apply init} = Tuple.rev apply |> tns |> init

                Tuple.foldr f dim finally {init apply=()}

            inl tns = Struct.map' (inl _ -> s.CudaTensor.create_view {dim elem_type=float}) d.init |> heap
            function
            | .data -> tns
            | .init -> Struct.iter2' (inl init tns -> init (init ()) tns) d.init tns
            | .save stream -> Struct.iter2' (inl f tns -> f stream tns s) d.save tns
            | .load stream -> Struct.iter2' (inl f tns -> f stream tns s) d.load tns
```

Here it is, the fragment of my dreams. Let me take a break here. Let me commit this.

As I've decided, I am not going to use a zipper representation for both the dimensions and the initializer as that would make it difficult to have separate initializers for different outer dimensions. I would be limited to only being able to specify initializers for the innermost dimension.

Doing it like this is the ideal.

I lazed too much in the morning, I only did this segment in the last half hour.

https://www.youtube.com/watch?v=bx5gJwD29p8&feature=youtu.be
TDLS: Bayesian Deep Learning on a Quantum Computer

I spent most of it watching this video and reading Baki. The video was interesting, but this last point is a confession from me. I'll try and do better from here on out. I will definitely get the LSTM done today. Let me get food and then I will resume after the break.