I've written a fairly insane amount of untested code today.
I really have so many doubts about the design.

There is just no helping it, though it is costing me a lot of time, I will just have to evolve it as I go along.

I might have made a serious error in separating primals and the adjoints from the tensor. Maybe I should create a specialized class just for dual numbers and have the learning module expect to be passed it.

...

A thought comes to me? Is Spiral's ML library badly designed? Yeah, probably. I should be doing less reflection and more message passing.

It is going to have to go through a large redesign if I ever want to open up the gates to higher order differentiation.

The specific problem I am thinking of is anywhere the `{block}` pattern is used. Occurences of `{block}` might be an antipattern.

...

Let me call it a day here.

There is no such thing as the perfect design. It is good to dwell on the true value of the code I am making from time to time.

Regardless of the library's fate, what it is giving me is truly top notch exercise, so it has great evolutionary value either way.

And I will keep evolving it until it gives me what I want.

It won't be long now until I move to the next step. Whether it is poker or the Singularity, the plan has always boiled do to 'use RL algorithms on games.' The devil lies in the execution.

Spiral is the seed of the future.

Once I am done with the 5 bullet points I will be able to move on to the next step. It will mark a milestone of me mastering first order differentiation and will allow me to start scaling up these algorithms.

Grasp a leaf and you will grasp hundred as long as you keep grasping.

I could have done poker in a few months back in 2015, instead it took me over ten times that time just to get to the entry point.

That won't change the fact that trivializing the optimal RNN compilation is a significant achievement whose benefits I will draw from down the line.

It makes sense to do this work now when the algorithms are still crap rather than in the 20s when sparks will be flying.