I got up a bit earlier. I still have no idea about the design - I might add a block to tree ranges in order to enable rewraping. I might also rename HostTensorView to just View.
But that is not important. As I said, I will evolve this thing as it goes along. Today the focus will be on finishing up the damn LSTM. After that will come the plastic RNN and K-FAC.

I want to get this done so I can move on to more interesting things. I'll do it perfectly later when I am done with poker.