Done lazing around and thanks to that my stress is low. It is time to deal with the LSTM. Let me get the first three points out of the way completely.
For poker I should not need much else on the NN side but K-FAC and the plastic RNN. If that is not enough to grasp it then I truly am lost and will have to consider a change of career.

The difficult parts that I will need Spiral for would be making the poker games, interfaces, asynchronous training and everything else along those lines. It won't be easy, but I will be able to do it.

It is time to start getting serious about things that will make me money. I can ket the ML community figure out metalearning while I do that.

I need to spend at least a year playing with games rather than languages and architectures.