The stupid thing finally compiles. Let me try running it.
```
The learning rate is 2 ** -11.00083
Starting timing for: iteration 0
The time was 00:00:05.4627241 for: iteration 0
Training: 3.07245799773099
The learning rate is 2 ** -11.00083
Starting timing for: iteration 1
The time was 00:00:05.4525358 for: iteration 1
Training: 2.39064590340508
The learning rate is 2 ** -11.00083
Starting timing for: iteration 2
The time was 00:00:05.3964010 for: iteration 2
Training: 2.19443693074316
The learning rate is 2 ** -11.00083
Starting timing for: iteration 3
The time was 00:00:05.3372302 for: iteration 3
Training: 2.06918151500662
The learning rate is 2 ** -11.00083
Starting timing for: iteration 4
The time was 00:00:05.4317452 for: iteration 4
Training: 1.98193556295626
```

This LSTM is quite fast. Let me try increasing the learning rate.

Nope, with -10 it diverges and with -10.5 it does on par with the above.

But regardless, this is quite good performance. The LSTM is now on par with the MI-RNN in terms of running time. I remember the old LSTM being over 3x slower.

Let me take a break here. I certainly earned it after the debugging session.

After I get back, I'll clean things up a little.