"4:20pm.
```
                weights =
                    state = weight_streams tanh (size, size)
                    input = weight_streams tanh (sublayer_size, size)
                    bias = {
                        si = const one (1,size)
                        i = const half (1,size)
                        s = const half (1,size)
                        c = zero (1,size)
                        }
                covariance =
                    state = covariance {in=size; out=size}
                    input = covariance {in=sublayer_size; out=size}
                    bias = {
                        si = covariance {out=size}
                        i = covariance {out=size}
                        s = covariance {out=size}
                        c = covariance {out=size}
                        }
```

I know how I will automate this. I will just add an extra pass to do this automatically.

I can fold the covariance in the weights somehow. But that is not the problem.

4:30pm.

```
                    inl bias = weights.bias
                    map CudaAD.generalized_mi_tanh {input state bias}
```

The damn biases are the problem.

4:40pm. Yeah, because of them I am not going to be able to do the fourth bullet point today.

But there is a way to do this. It is an idea that has occurred to me from time to time. I guess it is time I entertain it more seriously.

What I will do is allow singleton dimensions with sizes of 0. By sizes I mean the `size` used to calculate the offsets that resides in the `bodies`. This will allow me to apply them without increasing the offset.

4:45pm. Let me just do this thing and then I will call it a day. I'll leave getting everything else in place for tomorrow."