"5:20pm.
```
/// Expands the singular dimensions of a tensor to a specified size depending on the argument passed to it.
/// It does that without allocating or moving memory by setting the size for the offset calculation to zero.
inl expand_singular dim tns =
    inl dim = Tuple.wrap dim |> Tuple.map2 (inl cur dim -> match dim with () -> cur | _ -> dim) tns.dim
    inl is_expandable =
        Tuple.map2 (inl cur dim ->
            if cur <> dim then
                assert (cur = 1) "Only singular dimensions can be expanded."
                true
            else
                false
            ) tns.dim dim
    tns.update_body (inl d -> {d with size = Tuple.map2 (inl cur is_expandable -> if is_expandable then 0 else cur) self is_expandable})
       .set_dim dim
```

Let me do it like this. I really should support `()` as inputs to this function.

5:30pm.

```
    inl expand_singular dim tns s =
        inl out = Tensor.expand_singular dim (primal tns) |> dr s
        {
        out
        bck=met _ -> bck_add_bias (adjoint out) (adjoint tns) s
        }
```

Yeah, this is quite good. I like this. The principles of expansion are relatively simple, so I think I will be able to get away with this. As the expanded dimensions' offset sizes get set to zero that also makes them non-contiguous so this is not an unsafe operation."