Uagh, optimizing KFAC, uagh...
Please spare me. There is no really good way to describe the dependencies between operations in order to make it easier to decide. What I have here is simply the safe baseline.

...

Let me ignore those issues for a moment.

In order to finish the mi layer, what I need to do is also add the k and epsilon. For that I need to extend the Initializer once more. Let me do that. Also let me increment the counter in matmult and extend.