Ah, ok, I get it completely. I fucked up. The linear layer was not optimizing when I got that factorization error.
That solves that mystery. I solved this one precognitively it seems. I hope this starts a trend.

K-FAC actually works better now that I've changed the way I do updates to the covariance matrices. Nice.

...So what do I do next?

The LSTM.

But it is harder than it seems at first glance. I mean I could do it quick and dirty, but to do it properly I need to focus on the design.

I want to try an alternative MI layer that naturalized the biases together.

Let me do that one first. I might not have the energy to completely finish the 4th bullet point today, but that is fine. What is important that I do it properly.