``` The learning rate is 2 ** -12.00083 Starting timing for: iteration 0 The time was 00:00:16.1029788 for: iteration 0 Training: 2.01723659959594 The learning rate is 2 ** -12.00083 Starting timing for: iteration 1 The time was 00:00:16.0406749 for: iteration 1 Training: 1.56266987949322 The learning rate is 2 ** -12.00083 Starting timing for: iteration 2 The time was 00:00:16.0798919 for: iteration 2 Training: 1.48401048543575 The learning rate is 2 ** -12.00083 Starting timing for: iteration 3 The time was 00:00:16.0405355 for: iteration 3 Training: 1.43025916942519 The learning rate is 2 ** -12.00083 Starting timing for: iteration 4 The time was 00:00:16.0237230 for: iteration 4 Training: 1.39687958962339 ```
This is mi'' with a layer size of 384. Let me try the other one.

```
The learning rate is 2 ** -12.00083
Starting timing for: iteration 0
The time was 00:00:15.6904961 for: iteration 0
Training: 2.01424511492608
The learning rate is 2 ** -12.00083
Starting timing for: iteration 1
The time was 00:00:15.6647071 for: iteration 1
Training: 1.55568364463569
The learning rate is 2 ** -12.00083
Starting timing for: iteration 2
The time was 00:00:15.6825143 for: iteration 2
Training: 1.48161535599159
The learning rate is 2 ** -12.00083
Starting timing for: iteration 3
The time was 00:00:15.9925207 for: iteration 3
Training: 1.43097827509936
The learning rate is 2 ** -12.00083
Starting timing for: iteration 4
The time was 00:00:15.6698227 for: iteration 4
Training: 1.39722753063132
```

I think I am done for the day.