"10:40am.
```
The learning rate is 2 ** -12.00083
Starting timing for: iteration 0
The time was 00:00:07.8944512 for: iteration 0
Training: 2.99418740479193
The learning rate is 2 ** -12.00083
Starting timing for: iteration 1
The time was 00:00:07.9051468 for: iteration 1
Training: 2.28131307678653
The learning rate is 2 ** -12.00083
Starting timing for: iteration 2
The time was 00:00:07.8577200 for: iteration 2
Training: 2.07132536960382
The learning rate is 2 ** -12.00083
Starting timing for: iteration 3
The time was 00:00:07.9334225 for: iteration 3
Training: 1.95597803246175
The learning rate is 2 ** -12.00083
Starting timing for: iteration 4
The time was 00:00:07.8815988 for: iteration 4
Training: 1.86936240027631
```

Is this really right? Let me compare it with the standard LSTM with these learning rates.

10:45am.

```
The learning rate is 2 ** -12.00083
Starting timing for: iteration 0
The time was 00:00:06.2644232 for: iteration 0
Training: 3.23010642315699
The learning rate is 2 ** -12.00083
Starting timing for: iteration 1
The time was 00:00:06.2401139 for: iteration 1
Training: 2.61459843749317
The learning rate is 2 ** -12.00083
Starting timing for: iteration 2
The time was 00:00:06.2342508 for: iteration 2
Training: 2.36606166650788
The learning rate is 2 ** -12.00083
Starting timing for: iteration 3
The time was 00:00:06.3241627 for: iteration 3
Training: 2.24756699121858
The learning rate is 2 ** -12.00083
Starting timing for: iteration 4
The time was 00:00:06.2770964 for: iteration 4
Training: 2.15657621325569
```

The thing really got slower then yesterday. Is this the GPU being finicky or is it something else?

Also K-FAC is not getting particularly good results here. Let me make another layer in a standard fashion. I'll start with that. Then I'll go to the previous commit to test the speed."