``` The learning rate is 2 ** -12.00083 Starting timing for: iteration 0 The time was 00:00:10.1025500 for: iteration 0 Training: 2.42038412627024 The learning rate is 2 ** -12.00083 Starting timing for: iteration 1 The time was 00:00:10.0652308 for: iteration 1 Training: 1.87027217566912 The learning rate is 2 ** -12.00083 Starting timing for: iteration 2 The time was 00:00:10.0980017 for: iteration 2 Training: 1.7253689021283 The learning rate is 2 ** -12.00083 Starting timing for: iteration 3 The time was 00:00:10.0639450 for: iteration 3 Training: 1.64821506009834 The learning rate is 2 ** -12.00083 Starting timing for: iteration 4 The time was 00:00:10.1010982 for: iteration 4 Training: 1.60140345710234 ```
Yeah, the layer I made was not correct. This is exactly what I am afraid of with having to plug those covariances in manually in the optimize function. Bullshit.

I need to figure out why this is happening. Let me find the error and then I'll have my lesson from that.