I went back to that commit where I got 5.4s. per iteration This is what I get now.
```
The learning rate is 2 ** -11.00083
Starting timing for: iteration 0
The time was 00:00:06.0773493 for: iteration 0
Training: 3.07386548853452
The learning rate is 2 ** -11.00083
Starting timing for: iteration 1
The time was 00:00:05.9912423 for: iteration 1
Training: 2.38990912003004
The learning rate is 2 ** -11.00083
Starting timing for: iteration 2
The time was 00:00:06.2118176 for: iteration 2
Training: 2.18927248420796
The learning rate is 2 ** -11.00083
Starting timing for: iteration 3
The time was 00:00:06.1336637 for: iteration 3
Training: 2.07170364824945
The learning rate is 2 ** -11.00083
Starting timing for: iteration 4
The time was 00:00:06.2093166 for: iteration 4
Training: 1.98362402588485
```

Fuck GPUs.

There is no way that this could be due to anything on the host side.