``` The learning rate is 2 ** -12 At iteration 9 the cost and the sign accuracy are 26.25132 and 0.758/1. At iteration 19 the cost and the sign accuracy are 24.93219 and 0.862/1. At iteration 29 the cost and the sign accuracy are 23.46021 and 0.932/1. At iteration 39 the cost and the sign accuracy are 22.47039 and 0.97/1. At iteration 49 the cost and the sign accuracy are 21.22199 and 0.966/1. At iteration 59 the cost and the sign accuracy are 19.84457 and 0.982/1. At iteration 69 the cost and the sign accuracy are 18.80985 and 0.982/1. ```
Well, it is not quite exactly the same thing as last time, but it works. The differences might be due to numerical instability or something like that. Maybe I changed the learning rates? It does not matter. I am sure it works.

I think it is finally time to merge this branch with master.