https://arxiv.org/abs/1809.09147 Better Safe than Sorry - Evidence Accumulation Allows for Safe Reinforcement Learning
This paper could be interesting.

https://openreview.net/forum?id=H1lJJnR5Ym
Exploration by random distillation

"The TLDR would've been a better title: "A simple exploration bonus is introduced and achieves state of the art performance in 3 hard exploration Atari games.""

https://www.reddit.com/r/MachineLearning/comments/9ol1fb/r_piano_genie_an_intelligent_musical_interface/

Also an interesting thread.

"here's a section on the intuition. I might summarize it poorly, but basically, by learning to predict the output of another random network, the bigger the error, the more novel the state, and exploring novel states leads to success in these games. As the predictor gets better, the error goes down, and the state being observed therefore is less novel."

Here is a quote on that paper - it seems that they are using exactly the scheme mentioned in the IAC paper.

Let me spend my morning studying these papers.

Then I will start programming. I am starting to get motivated for this.