Done with reading. Let me do some programming.
The two papers were quite interesting.

"By observing the RND agent (goo.gl/DGPC8E), we notice that frequently once it obtains all the extrinsic rewards that it knows how to obtain reliably (as judged by the extrinsic value function), the agent settles into a pattern of behavior where it keeps interacting with potentially dangerous objects. For instance in Montezuma’s Revenge the agent jumps back and forth over a moving skull, moves in between laser gates, and gets on and off disappearing bridges. We also observe similar behavior in Pitfall!. It might be related to the very fact that such dangerous states are difficult to achieve, and hence are rarely represented in agent’s past experience compared to safer states."

I made a comment on the RL sub that this is something I tend to do when I am bored in games, or at least, used to do back when I was a kid.

The accumulator paper is something I will be keeping in mind and waiting for follow ups. I wonder if plastic RNNs would do better on the task covered in that paper.

Well, that does not matter.

Forget that stuff from here on out.

I've stoked up on despair and hatred, and I am ready to carve it into code. It is time to slowly get into it.

Strangely, I've been feeling up for it over the night, but now that it is time to do serious work I feel like lazing.

Right now I am thinking of what I should do and I think I'll go for a complete redesign of the plastic RNN from the ground up. It is honestly too complicated to do it by small changes.

Once I fused the multiplies that would wreck everything anyway.

Let me start by sketching out the solution.