``` inl bias, input, state = wrap_split ((), dim.inner) (weights.bias.weight, input, state) ```
I renamed it to `wrap_split`. Actually, it is not surprising that this works. It is quite a straightforward deduction to prove that it would work. The amazing thing is that I came to the conclusion that this would not work yesterday. That is why writing this out feels so strange to me right now. Let me convert the rest too.