"4:10pm. Ohhh, it compiles. Let me try running it.
```
The learning rate is 2 ** -11.5
At iteration 9 the cost and the sign accuracy are 25.46727 and 0.728/1.
At iteration 19 the cost and the sign accuracy are 25.16917 and 0.744/1.
At iteration 29 the cost and the sign accuracy are 24.84167 and 0.758/1.
At iteration 39 the cost and the sign accuracy are 25.24682 and 0.748/1.
At iteration 49 the cost and the sign accuracy are 25.46949 and 0.716/1.
At iteration 59 the cost and the sign accuracy are 25.12858 and 0.752/1.
At iteration 69 the cost and the sign accuracy are 25.19061 and 0.74/1.
At iteration 79 the cost and the sign accuracy are 25.04963 and 0.764/1.
````

It is definitely much faster, but unfortunately it does not work correctly.

4:20pm.

```
                    {
                    input=s.CudaTensor.zero {elem_type=float; dim=weights.sublayer_size, size}
                    state=s.CudaTensor.zero {elem_type=float; dim=size, size}
                    }

            inl state =
                match d with
                | {state={state}} -> state
                | _ -> s.CudaTensor.zero {elem_type=float; dim=span, size}
```

Ah, it has to be this. Every time an error like this happens, I start thinking whether I missed a sign somewhere, but now that the implementations are made of iron, I should know to suspect the bigger things. I forgot to wrap the numbers in a dual here. Let me try it again...

...No wait, that can't be it. This is right.

4:25pm. I was too cocky. I had been going without runtime bugs for a while now that I started to hope that I was done with them for good. I am not.

Thus bug is going to be hard to find.

I am going to have to think a little about where I even want to begin the search.

Let me start with sanity checking."