``` inl set_adjoint x out =     match x with     | {adjoint} -> adjoint 0 <- out ()     | _ -> () ```
Oh, oh crap. It is really setting the adjoint rather than adding to it. You are shitting me?

This is definitely the error.

Edit:

```
The learning rate is 2 ** -11.5
At iteration 0 the cost and the sign accuracy are 10.61978 and 0.75/1.
At iteration 1 the cost and the sign accuracy are 10.38275 and 0.8/1.
At iteration 2 the cost and the sign accuracy are 10.52077 and 0.8/1.
At iteration 3 the cost and the sign accuracy are 10.61555 and 0.8/1.
At iteration 4 the cost and the sign accuracy are 10.51251 and 0.9/1.
At iteration 5 the cost and the sign accuracy are 10.70124 and 0.75/1.
At iteration 6 the cost and the sign accuracy are 10.71004 and 0.65/1.
At iteration 7 the cost and the sign accuracy are 10.80053 and 0.7/1.
At iteration 8 the cost and the sign accuracy are 10.68788 and 0.7/1.
At iteration 9 the cost and the sign accuracy are 10.31518 and 0.85/1.
At iteration 10 the cost and the sign accuracy are 10.84016 and 0.6/1.
```

Or maybe not.

It is true that this was a significant error, but it not the error.