6:55pm. I can't figure out what is wrong with `plastic_rnn''`. Let me make one last experiment for the day. I will just replace the `broadcasting_activation` and the final `activation` in the old plastic RNN. That will tell me for certain whether the error is in CudaAD somewhere.
7pm.

```
inl plastic_rnn_cell {static alpha plastic} = static + alpha * plastic
inl plastic_rnn_cell' {input state bias} = input + state + bias |> tanh
```

Let me add these two.

7pm.

```
At iteration 37 the cost and the sign accuracy are 7.445261 and 1/1.
At iteration 38 the cost and the sign accuracy are 7.432314 and 1/1.
At iteration 39 the cost and the sign accuracy are 9.387165 and 0.75/1.
At iteration 40 the cost and the sign accuracy are 8.571859 and 0.85/1.
At iteration 41 the cost and the sign accuracy are 7.100524 and 1/1.
At iteration 42 the cost and the sign accuracy are 6.518993 and 1/1.
At iteration 43 the cost and the sign accuracy are 9.596906 and 0.8/1.
At iteration 44 the cost and the sign accuracy are 7.941875 and 1/1.
At iteration 45 the cost and the sign accuracy are 7.459081 and 0.9/1.
At iteration 46 the cost and the sign accuracy are 4.498063 and 1/1.
At iteration 47 the cost and the sign accuracy are 7.318504 and 0.95/1.
At iteration 48 the cost and the sign accuracy are 6.100853 and 1/1.
At iteration 49 the cost and the sign accuracy are 5.697697 and 1/1.
```

Just like the standard one. Let me replace the Oja update too.

```
At iteration 38 the cost and the sign accuracy are 10.19382 and 0.85/1.
At iteration 39 the cost and the sign accuracy are 10.64829 and 0.7/1.
At iteration 40 the cost and the sign accuracy are 10.76359 and 0.7/1.
At iteration 41 the cost and the sign accuracy are 10.45262 and 0.7/1.
At iteration 42 the cost and the sign accuracy are 9.899019 and 0.85/1.
At iteration 43 the cost and the sign accuracy are 10.44715 and 0.8/1.
At iteration 44 the cost and the sign accuracy are 10.7695 and 0.7/1.
At iteration 45 the cost and the sign accuracy are 10.59374 and 0.75/1.
At iteration 46 the cost and the sign accuracy are 10.39544 and 0.75/1.
At iteration 47 the cost and the sign accuracy are 9.975306 and 0.9/1.
At iteration 48 the cost and the sign accuracy are 10.84893 and 0.8/1.
At iteration 49 the cost and the sign accuracy are 10.71014 and 0.75/1.
```

So it was in the Oja update.

I have you now, I have you!