```                 inm H =                     inm input = map (CudaAD.oja_update n) {out H=H.input; input}                     inm state = map (CudaAD.oja_update n) {out H=H.state; input=state}                     succ {input state}
                succ {out state={state=out; H}}
```

It is not a matter of separating them, it just does not work. The error is here 100%.

```
H + n * (input * out - out * out * H)
```

The error is literally within this simple expression. But what should I even suspect here? The `-`?

Let me fiddle around with it for a little. At this point, I am really dead tired, but I will do it.