10:20am. The gradients are clearly different. Let me set the pattern size to 2.
```
[|
    [|0; 0|]
    [|0; 0|]
|]

[|
    [|0; 0|]
    [|-0.01922576; 0.0002036725|]
|]

[|
    [|-1.339049E-05; 3.960411E-07|]
    [|-0.01808944; 0.0001914606|]
|]

[|
    [|-9.70451E-06; -1.35785E-06|]
    [|-0.01704319; 0.0001822964|]
|]

[|
    [|-3.829488E-05; -4.894906E-06|]
    [|-0.01608344; 0.0001679011|]
|]
```

This one is correct.

```
[|
    [|0; 0|]
    [|0; 0|]
|]

[|
    [|0; 0|]
    [|-0.01908011; 0.0002036209|]
|]

[|
    [|-8.340304E-05; 4.094784E-07|]
    [|-0.01788592; 0.0001913888|]
|]

[|
    [|-8.835777E-05; -1.317391E-06|]
    [|-0.01683481; 0.000182218|]
|]

[|
    [|-0.000108291; -4.835509E-06|]
    [|-0.01588479; 0.0001678474|]
|]
```

I have no idea. I am just printing this stuff in order to juggle my mind. How should I proceed from here?

This is going to take some effort, but I guess I need to split the Oja update into pieces and try to isolate the error.

It is going to be difficult to do this for the template, but for the map it won't be hard. Let me do it. There is definitely some error on broadcasting_link or something like that, but I just cannot see it.

...Wait...

An idea just came to my mind.